+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Types of changes:
 
 * Handle where getting text snippets returned True rather than skipping
 * Fix 'Smoking status: never assessed'
+* Fix hypothetical/questions sentences
 
 ## [0.0.1] - 2024-01-12